Added the single data transfer function.
I have added the single data transfer function which either stores data from the source register into the memory or loads data from the memory into the destination reg, determined by the instruction.
The instruction also determines the offset of the base register and whether the offset is immediate or shifted register and whether it is added or subtracted from the base reg.